+++ essential-real-estate/includes/widgets/acf/g5plus-acf.php
@@ -31,12 +31,12 @@ if (!class_exists('GF_Acf_Widget_Fields')) {
             $field_values = is_array($this->value) && array_key_exists('fields', $this->value) ? $this->value['fields'] : '';
             $extra_values = is_array($this->value) && array_key_exists('extra', $this->value) ? $this->value['extra'] : '';
             $x = 0;
-	        $plugin_path = untrailingslashit(plugin_dir_path(__FILE__));
             $is_edit_mode = (isset($field_values) && is_array($field_values) && count($field_values) > 0) || (isset($extra_values) && is_array($extra_values) && count($extra_values) > 0);
+            $allowed_fields = array('checkbox','icon','image','number','select','text','text-area');
 	        if ($is_edit_mode) {
-		        include($plugin_path . '/templates/edit.php');
+                include (ERE_PLUGIN_DIR . '/includes/widgets/acf/templates/edit.php');
 	        } else {
-		        include($plugin_path . '/templates/new.php');
+                include (ERE_PLUGIN_DIR . '/includes/widgets/acf/templates/new.php');
 	        }
         }
     }