+++ essential-real-estate/includes/widgets/acf/templates/checkbox.php
@@ -1,3 +1,18 @@
+<?php
+/**
+ * @var $require_element
+ * @var $require_element_id
+ * @var $require_compare
+ * @var $require_values
+ * @var $field_output_name
+ * @var $field_title
+ * @var $field_output_id
+ * @var $field_value
+ */
+if (!defined('ABSPATH')) {
+    exit; // Exit if accessed directly
+}
+?>
 <div class="checkbox-wrap" data-require-element="<?php if(isset($require_element)){ echo esc_attr($require_element);} ?>"
      data-require-element-id="<?php if(isset($require_element_id)){ echo esc_attr($require_element_id);} ?>"
      data-require-compare="<?php if(isset($require_compare)){ echo esc_attr($require_compare);} ?>"